improved factorization of tests
